Tax rates are calulated from the DB

+++ Code/TaxCalculator V1.1/TaxCalculator/Models/TaxPayerSalaryDetails.cs
@@ -7,7 +7,7 @@ namespace TaxCalculator.Models
 {
   public partial class TaxPayer
   {
-    public float YearlySalary { get; set; }
+    public decimal YearlySalary { get; set; }
     public List<Deduction> YearlyDeductions { get; set; }
   }
 }
